Move more shaders related codes away from material.cpp. #122
